refactor: 포스트와 추천의 양방향 의존성을 단방향으로 수정 (#194)
* refactor: PostService와 ThumbsService의 양방향 연관을 중간 interface를 두어 thumbs -> post만 되도록 수정함

* Feat/#173 조회수 기록 (#186)

* test: 게시글의 조회수 증가와 관련된 서비스 기능을 tdd로 테스트

* reafactor: memberIdDto를 DTO에서 VO로 격상시키고,
로그인 한 사용자 id를 나타내는 MemberId와
로그인 하지 않은 사용자 id를 나타내는 AnonymousMemberId로 분리

* feat: 게시글 상세조회 시 조회수를 증가시키는 기능 구현

* fix: 테스트 케이스는 dev 환경에서 동작하도록 개선

* test: 게시글의 조회수 증가와 관련된 테스트 케이스를 ATDD로 작성

* test: 기능 구현으로 인해 깨지는 레거시 테스트 교정

* fix: 깨지는 컴파일 개선

* refactor: MemberId 추상 클래스 하위에 ActiveMemberId와 AnonymousMemberId를 두어 상속관계를 명확히 개선

* refactor: Post의 조회수와 관련된 인수 테스트를 별도의 클래스로 분리

* refactor: 불필요한 박싱 언박싱 비용을 줄이기 위해 viewCount를 원시값으로 변경

* refactor: 변수에 ActiveMemberId 대신 상위 객체인 MemberId를 사용

* refactor: 불필요한 DDL 변경을 롤백

* style: 컨벤션에 맞게 코드 수정

* Feat/#135 그래프 (#171)

* feat: 라인 2개가 그려지는 그래프 구현

- MPAndroidChart 적용

* feat: 더미 데이터를 view model의 live data 이용

- 두 연/월 사이에 있는 날짜들을 구하는 로직은 도메인에서 진행

* feat: 그래프 클릭 시 해당 포인트의 가격 값 보여주도록 구현

- MarkerView 이용

* refactor: 그래프를 별도의 클래스로 분리 및 주석 추가

* refactor: 클래스 이름 변경

* Refactor/#161 검색 기능 개선 (#189)

* docs: profile 설정용 properties 깃 이그노어 적용

* feat: 띄어쓰기를 기준으로 다중 키워드로 검색 가능하도록 구현

* feat: 대소문자 상관없이 검색되도록 구현

* refactor: post의 조회와 cud 작업을 controller와 service레이어에서 각각 분리함

* refactor: 페이징 된 게시글 목록을 Slice타입 객체로 response를 내려주도록 수정함

* test: 테스트 깨지는 것 해결함

* test: 테스트 케이스 추가

* fix: merge 중에 난 conflict 해결함

* fix: merge 중에 난 conflict 해결함2

* refactor: 호이 리뷰 부분적으로 반영함

* refactor: PostService 클래스 레벨의 transactional 제거함

* refactor: PostService와 ThumbsService의 양방향 연관을 중간 interface를 두어 thumbs -> post만 되도록 수정함

* fix: merge중 conflict 난 것 해결

* refactor: 이리내, 여우 피드백 반영
