+++ backend/src/main/java/edonymyeon/backend/post/application/PostReadService.java
@@ -10,9 +10,6 @@ import edonymyeon.backend.post.application.dto.*;
 import edonymyeon.backend.post.domain.Post;
 import edonymyeon.backend.post.repository.PostRepository;
 import edonymyeon.backend.post.repository.PostSpecification;
-import edonymyeon.backend.thumbs.application.ThumbsService;
-import edonymyeon.backend.thumbs.dto.AllThumbsInPostResponse;
-import edonymyeon.backend.thumbs.dto.ThumbsStatusInPostResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
@@ -35,7 +32,7 @@ public class PostReadService {
 
     private final MemberRepository memberRepository;
 
-    private final ThumbsService thumbsService;
+    private final PostThumbsService thumbsService;
 
     private final Domain domain;
